Refactor: remove some methods in tests to upper level to reduce duplication
