+++ integ-test/src/test/java/org/opensearch/sql/legacy/SQLIntegTestCase.java
@@ -6,7 +6,6 @@
 package org.opensearch.sql.legacy;
 
 import static com.google.common.base.Strings.isNullOrEmpty;
-import static org.opensearch.sql.common.setting.Settings.Key.SCRIPT_CONTEXT_MAX_COMPILATIONS_RATE_PATTERN;
 import static org.opensearch.sql.legacy.TestUtils.createIndexByRestClient;
 import static org.opensearch.sql.legacy.TestUtils.getAccountIndexMapping;
 import static org.opensearch.sql.legacy.TestUtils.getAliasIndexMapping;
@@ -64,11 +63,8 @@ import java.nio.charset.StandardCharsets;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.Paths;
-import java.util.ArrayList;
-import java.util.List;
 import java.util.Locale;
 import java.util.Map;
-import java.util.Objects;
 import java.util.concurrent.Callable;
 import javax.management.MBeanServerInvocationHandler;
 import javax.management.ObjectName;
@@ -220,46 +216,6 @@ public abstract class SQLIntegTestCase extends OpenSearchSQLRestTestCase {
   protected void resetMaxResultWindow(String indexName) throws IOException {
     updateIndexSettings(
         indexName, "{ \"index\": { \"max_result_window\": " + DEFAULT_MAX_RESULT_WINDOW + " } }");
-  }
-
-  /**
-   * Increases the maximum script compilation rate for all script contexts for tests. This method
-   * sets an unlimited compilation rate for each script context when the
-   * script.disable_max_compilations_rate setting is not enabled, allowing tests to run without
-   * hitting compilation rate limits.
-   *
-   * @throws IOException if there is an error retrieving cluster settings or updating them
-   */
-  protected void increaseMaxCompilationsRate() throws IOException {
-    // When script.disable_max_compilations_rate is set, custom context compilation rates cannot be
-    // set
-    if (!Objects.equals(
-        getClusterSetting(
-            Settings.Key.SCRIPT_DISABLE_MAX_COMPILATIONS_RATE.getKeyValue(), "persistent"),
-        "true")) {
-      List<String> contexts = getScriptContexts();
-      for (String context : contexts) {
-        String contextCompilationsRate =
-            SCRIPT_CONTEXT_MAX_COMPILATIONS_RATE_PATTERN.getKeyValue().replace("*", context);
-        updateClusterSettings(
-            new ClusterSetting("persistent", contextCompilationsRate, "unlimited"));
-      }
-    }
-  }
-
-  protected List<String> getScriptContexts() throws IOException {
-    Request request = new Request("GET", "/_script_context");
-    Response response = client().performRequest(request);
-    String responseBody = getResponseBody(response);
-    JSONObject jsonResponse = new JSONObject(responseBody);
-    JSONArray contexts = jsonResponse.getJSONArray("contexts");
-    List<String> contextNames = new ArrayList<>();
-    for (int i = 0; i < contexts.length(); i++) {
-      JSONObject context = contexts.getJSONObject(i);
-      String contextName = context.getString("name");
-      contextNames.add(contextName);
-    }
-    return contextNames;
   }
 
   /** Provide for each test to load test index, data and other setup work */
@@ -442,17 +398,6 @@ public abstract class SQLIntegTestCase extends OpenSearchSQLRestTestCase {
     return executeRequest(sqlRequest);
   }
 
-  protected static String executeRequest(final Request request, RestClient client)
-      throws IOException {
-    Response response = client.performRequest(request);
-    Assert.assertEquals(200, response.getStatusLine().getStatusCode());
-    return getResponseBody(response);
-  }
-
-  protected static String executeRequest(final Request request) throws IOException {
-    return executeRequest(request, client());
-  }
-
   protected JSONObject executeQueryWithGetRequest(final String sqlQuery) throws IOException {
 
     final Request request = buildGetEndpointRequest(sqlQuery);
@@ -487,24 +432,6 @@ public abstract class SQLIntegTestCase extends OpenSearchSQLRestTestCase {
 
   protected static JSONObject updateClusterSettings(ClusterSetting setting) throws IOException {
     return updateClusterSettings(setting, client());
-  }
-
-  protected static JSONObject getAllClusterSettings() throws IOException {
-    Request request = new Request("GET", "/_cluster/settings?flat_settings&include_defaults");
-    RequestOptions.Builder restOptionsBuilder = RequestOptions.DEFAULT.toBuilder();
-    restOptionsBuilder.addHeader("Content-Type", "application/json");
-    request.setOptions(restOptionsBuilder);
-    return new JSONObject(executeRequest(request));
-  }
-
-  protected static String getClusterSetting(String settingPath, String type) throws IOException {
-    JSONObject settings = getAllClusterSettings();
-    String value = settings.optJSONObject(type).optString(settingPath);
-    if (StringUtils.isEmpty(value)) {
-      return settings.optJSONObject("defaults").optString(settingPath);
-    } else {
-      return value;
-    }
   }
 
   protected static class ClusterSetting {
